+++ openpecha/alignment/__init__.py
@@ -1,0 +1,10 @@
+
+class Alignment:
+  def __init__(self, id=None):
+    self.id = id
+  
+  def generate_segment_pairs(self):
+    pass
+  
+  def export(self, format=None):
+    pass