+++ openpecha/core/annotation.py
@@ -1,6 +1,6 @@
-from typing import Optional
+from typing import Dict, Optional
 
-from pydantic import BaseModel, FilePath, validator
+from pydantic import BaseModel, validator
 
 
 class Span(BaseModel):
@@ -22,14 +22,11 @@ class Span(BaseModel):
 
 class AnnBase(BaseModel):
     span: Span
-
-
-class VerseTypeAnn(AnnBase):
-    is_verse: bool = False  # whether ann in verse format
+    metadata: Dict = {}
 
 
 class Page(AnnBase):
-    page_info: Optional[str] = None # page payload
+    page_info: Optional[str] = None  # page payload
     imgnum: Optional[
         int
     ] = None  # image sequence no. from bdrc api, http://iiifpres.bdrc.io/il/v:bdr:I0888