+++ openpecha/serializers/__init__.py
@@ -1,3 +1,4 @@
+from .docx import DocxSerializer
 from .editor import EditorSerializer
 from .epub import EpubSerializer
 from .hfml import HFMLSerializer