+++ tests/serializers/test_epub_serializer.py
@@ -10,7 +10,7 @@ def test_opf_2_tsadra_serializer():
     with tempfile.TemporaryDirectory() as tmpdirname:
         serializer = EpubSerializer(opf_path)
         out_fn = serializer.serialize(output_path=tmpdirname)
-        assert Path(out_fn).name == "P000801.epub"
+        assert out_fn.name == "P000801.epub"
 
 
 def test_opf_2_html_serializer():
@@ -37,4 +37,3 @@ def test_opf_2_html_serializer():
         )
         serialized_html += f"{result}{footnote_ref_tag}</body>\n</html>"
         assert expected_serialized_html == serialized_html
-